finished/integrate-services/service-bus/python/client/pyproject.toml
[project]
name = "service_bus_messaging"
version="0.1.0"
description = "Process messages with Azure Service Bus queues, topics, and subscriptions"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.21.0",
    "azure-servicebus>=7.14.1",
]
